serving_pipeline/backend/requirements.txt
redis==5.2.1
fastapi==0.115.12
pydantic==2.11.2
uvicorn==0.34.0
pydantic-settings==2.8.1
python-multipart==0.0.20
torch==2.6.0
numpy==2.1.1
pillow==11.1.0
transformers[torch]==4.50.3
diffusers==0.32.2
ultralytics==8.3.101
python-dotenv==1.0.0